feat: add address book fetch fail page
